+++ source/pipes/math/date-time-latest.test.ts
@@ -1,0 +1,27 @@
+import { DateTime } from "luxon";
+
+import { DateTimeLatestPipe } from "./date-time-latest";
+
+describe("DateTimeLatestPipe", () => {
+
+  let pipe: DateTimeLatestPipe;
+
+  beforeEach(() => {
+    pipe = new DateTimeLatestPipe();
+  });
+
+  describe("#transform", () => {
+
+    it("transforms a list of DateTimes into the maximum of that list", () => {
+      const result = pipe.transform([
+        DateTime.fromISO("2006-01-02T15:04:05-07:00"),
+        DateTime.fromISO("2007-01-02T15:04:05-07:00"),
+        DateTime.fromISO("2005-01-02T15:04:05-07:00")
+      ]);
+
+      expect(result.year).toBe(2007);
+    });
+
+  });
+
+});